add files upload layoud and functionality, deleting, pdf creating

+++ controllers/admin.js
@@ -1,5 +1,6 @@
 const mongoose = require('mongoose');
 const Product = require('../models/product');
+const fileHelper = require('../util/file');
 const { validationResult } = require('express-validator');
 
 exports.getAddProduct = (req, res, next) => {
@@ -21,9 +22,24 @@ exports.getAddProduct = (req, res, next) => {
 exports.postAddProduct = (req, res, next) => {
   const title = req.body.title;
   const price = req.body.price;
+  const image = req.file;
   const description = req.body.description;
-  const imageUrl = req.body.imageUrl;
   const errors = validationResult(req);
+
+  if (!image) {
+    return res.status(422).render('admin/edit-product', {
+      pageTitle: 'Add Product',
+      path: '/admin/add-product',
+      editing: false,
+      hasError: true,
+      product: {
+        title,
+        price,
+        description
+      },
+      errorMessage: 'Attached file is not and image',
+    });
+  }
 
   if (!errors.isEmpty()) {
     console.log('errors.Array()', errors.array());
@@ -34,13 +50,14 @@ exports.postAddProduct = (req, res, next) => {
       hasError: true,
       product: {
         title,
-        imageUrl,
         price,
         description
       },
-      errorMessage: errors.array()[0].msg,
+      errorMessage: 'Attached file is not an image.',
     });
   }
+
+  const imageUrl = image.path;
 
   const product = new Product(
     {
@@ -52,14 +69,12 @@ exports.postAddProduct = (req, res, next) => {
       userId: req.user
     }
     );
-   
   product.save()
     .then(() => {
       console.log('Product created!');
       res.redirect('/admin/products');
     })
     .catch(err => {
-      console.log('AN ERROR OCCURED: ')
       console.log(err, 'err')
       // return res.status(500).render('admin/add-product', {
       //   pageTitle: 'Add Product',
@@ -91,7 +106,6 @@ exports.getEditProduct = (req, res, next) => {
   const prodId = req.params.productId;
   Product.findById(prodId)
     .then(product => {
-      console.log('Edit from admin: ', product);
       if (!product) {
         return res.redirect('/');
       }
@@ -116,7 +130,7 @@ exports.postEditProduct = (req, res, next) => {
   const prodId = req.body.productId;
   const updatedTitle = req.body.title;
   const updatedPrice = req.body.price;
-  const updatedImageUrl = req.body.imageUrl;
+  const image = req.file;
   const updatedDesc = req.body.description;
   const errors = validationResult(req);
 
@@ -128,7 +142,6 @@ exports.postEditProduct = (req, res, next) => {
       hasError: true,
       product: {
         title: updatedTitle,
-        imageUrl: updatedImageUrl,
         price: updatedPrice,
         description: updatedDesc,
         _id: prodId
@@ -149,7 +162,10 @@ exports.postEditProduct = (req, res, next) => {
         product.title = updatedTitle;
         product.price = updatedPrice;
         product.description = updatedDesc;
-        product.imageUrl = updatedImageUrl;
+        if (image) {
+          fileHelper.deleteFile(product.imageUrl);
+          product.imageUrl = image.path;
+        }
         return product.save()
           .then(updtProduct => {
             console.log('Updated product', updtProduct)
@@ -201,14 +217,20 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.deleteOne({_id: prodId, userId: req.user._id})
+  Product.findById(prodId)
+    .then((product) => {
+      if (!product) {
+        return next(new Error('Product not found'));
+      }
+      fileHelper(product.imageUrl);
+      return Product.deleteOne({_id: prodId, userId: req.user._id})
+    })
     .then(() => {
       res.redirect('/admin/products');
     })
     .catch(err => {
-      console.warn('err', err);
-        const error = new Error('Error')
-        error.httpStatusCode = 500;
-        return next(error);
-    });
+      const error = new Error('Error')
+      error.httpStatusCode = 500;
+      return next(error);
+    })
   };
